Rename BaseVOTest to GetSetReflectBaseTest.
