+++ soul-admin/src/test/java/org/dromara/soul/admin/query/SoulDictQueryTestGetSetReflect.java
@@ -15,19 +15,19 @@
  * limitations under the License.
  */
 
-package org.dromara.soul.admin.vo;
+package org.dromara.soul.admin.query;
 
-import org.dromara.soul.admin.vo.base.BaseVOTest;
+import org.dromara.soul.admin.vo.base.GetSetReflectBaseTest;
 
 /**
- * Test case for AuthPathVO.
+ * Test case for SoulDictQuery.
  *
- * @author midnight2104
+ * @author wuudongdong
  */
-public class AuthPathVOTest extends BaseVOTest<AuthPathVO> {
+public final class SoulDictQueryTestGetSetReflect extends GetSetReflectBaseTest<SoulDictQuery> {
 
     @Override
-    protected AuthPathVO getTargetClass() {
-        return new AuthPathVO();
+    protected SoulDictQuery getTargetClass() {
+        return new SoulDictQuery();
     }
 }
