+++ soul-admin/src/test/java/org/dromara/soul/admin/vo/RuleTestGetSetReflect.java
@@ -17,17 +17,17 @@
 
 package org.dromara.soul.admin.vo;
 
-import org.dromara.soul.admin.vo.base.BaseVOTest;
+import org.dromara.soul.admin.vo.base.GetSetReflectBaseTest;
 
 /**
- * Test case for DashboardUserVO.
+ * Test case for RuleVO.
  *
  * @author midnight2104
  */
-public class DashboardUserVOTest extends BaseVOTest<DashboardUserVO> {
+public class RuleTestGetSetReflect extends GetSetReflectBaseTest<RuleVO> {
 
     @Override
-    protected DashboardUserVO getTargetClass() {
-        return new DashboardUserVO();
+    protected RuleVO getTargetClass() {
+        return new RuleVO();
     }
 }
